backend/fastapi/requirements.txt
fastapi
uvicorn
librosa
soundfile
numpy
scipy
httpx
python-dotenv
python-multipart
aiofiles
openai>=1.0.0
huggingface_hub
pyannote.audio
